feat: Timer Log Entity를 Timer Entity로 변경 및 TimerRepository 인터페이스 및 구현체 구현
