--- client/src/entities/timer-log/api/types.ts
@@ -1,14 +1,0 @@
-import type { TimerLog } from "../model/types";
-
-export interface TimerLogDto {
-  id: string;
-  participantId: string;
-  value: number;
-  type: "start" | "stop" | "adjust";
-  createdAt: string;
-}
-
-export interface TimerLogRepository {
-  // Implement methods when api is ready
-  getTimerLogs(participantId: string): Promise<TimerLog[]>;
-}
