+++ client/src/entities/timer/lib/__test__/selectors.test.ts
