+++ client/src/entities/timer/lib/parse-dto.ts
