+++ client/src/entities/timer/model/__test__/reducers.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
 import { integrateLogs } from "../reducers";
 
-import type { TimerLog } from "@/entities/timer-log";
+import type { TimerLog } from "@/entities/timer";
 
 describe("integrateLogs", () => {
   const base = 1_000_000;
