+++ client/src/entities/timer/model/reducers.ts
