+++ client/src/entities/timer/model/types.ts
