+++ client/src/entities/timer/ui/formatters.ts
