+++ client/src/features/progress/model/types.ts
@@ -1,7 +1,7 @@
 import type { Competition } from "@/entities/competition";
 import type { Division } from "@/entities/division";
 import type { Participant } from "@/entities/participant";
-import type { TimerLog } from "@/entities/timer-log";
+import type { TimerLog } from "@/entities/timer";
 import type { Record } from "@/entities/record";
 import type { ManualRecord } from "@/entities/manual-record";
